refactor(@sounisi5011/encrypted-archive): move protobuf code into `src/protocol-buffers` directory
