+++ packages/encrypted-archive/docs/encrypted-archive-structure.md
@@ -3,7 +3,7 @@
 [unsigned varint]: https://github.com/multiformats/unsigned-varint
 [multicodec spec]: https://github.com/multiformats/multicodec
 [Protocol Buffers]: https://developers.google.com/protocol-buffers/
-[`src/core/header/protocol-buffers/header.proto`]: ../src/core/header/protocol-buffers/header.proto
+[`src/protocol-buffers/header.proto`]: ../src/protocol-buffers/header.proto
 
 Encrypted archive data consists of one "first chunk", followed by zero or more "subsequent chunks".
 By being divided into multiple chunks, encryption and decryption can be performed without having to expand all the huge data into memory.
@@ -31,7 +31,7 @@ The "first chunk" is at the beginning of the data. This chunk contains the follo
 3. Header
 
     Binary data encoded by [Protocol Buffers] that contains the data required for decryption.
-    The definition is the `Header` message in [`src/core/header/protocol-buffers/header.proto`] file.
+    The definition is the `Header` message in [`src/protocol-buffers/header.proto`] file.
 
 4. Ciphertext Length
 
@@ -59,7 +59,7 @@ The "subsequent chunk" exists after the "first chunk". This chunk contains the f
 2. SimpleHeader
 
     Binary data encoded by [Protocol Buffers] that contains the data required for decryption.
-    The definition is the `SimpleHeader` message in [`src/core/header/protocol-buffers/header.proto`] file.
+    The definition is the `SimpleHeader` message in [`src/protocol-buffers/header.proto`] file.
     It does not contain any data that overlaps with the Header. This makes the data length shorter.
 
 3. Ciphertext Length
