+++ packages/encrypted-archive/src/protocol-buffers/header_pb.d.ts
